requirements/base.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
django-filter==25.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.4
pytest-django==4.10.0
referencing==0.36.2
rpds-py==0.22.3
stripe==11.5.0
sqlparse==0.5.3
uritemplate==4.1.1
